[refactor] FCM 예외처리 리팩토링
