--- backend/src/main/java/com/coffee/backend/domain/auth/exception/LoginIdAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.coffee.backend.domain.auth.exception;
-
-public class LoginIdAlreadyExistsException extends RuntimeException {
-    public LoginIdAlreadyExistsException(String message) {
-        super(message);
-    }
-}
